feat: 국가/공항 디렉터리 시스템 추가 및 trip 수정 API 제거
- 국토교통부 세계공항 데이터 기반 국가/공항 디렉터리 구현
- Country/Airport 모델 및 마이그레이션 추가
- /v1/countries, /v1/airports API 엔드포인트 추가
- 공항 데이터 동기화 스크립트 (sync_airports.py) 추가
- airport_lookup 서비스를 DB 기반으로 전환 (fallback 지원)
- TripService에서 존재하지 않는 공항 코드 검증 강화
- SQLite 테스트 지원을 위한 BIGINT 타입 래퍼 추가
- trip 수정(PATCH) 엔드포인트 제거
- 마이그레이션 브랜치 병합 (b3c4d5e6f7g8 + 7c2bb0e9d2d3)

+++ app/db/models/user.py
@@ -1,16 +1,17 @@
 from datetime import datetime
 from typing import Any
 
-from sqlalchemy import TIMESTAMP, String, BigInteger, JSON, func
+from sqlalchemy import TIMESTAMP, String, JSON, func
 from sqlalchemy.orm import Mapped, mapped_column
 
 from app.db.base import Base
+from app.db.types import BIGINT
 
 
 class User(Base):
     __tablename__ = "users"
 
-    user_id: Mapped[int] = mapped_column(BigInteger, primary_key=True, autoincrement=True)
+    user_id: Mapped[int] = mapped_column(BIGINT, primary_key=True, autoincrement=True)
     device_uuid: Mapped[str | None] = mapped_column(String(36), unique=True)
     name: Mapped[str | None] = mapped_column(String(100))
     locale: Mapped[str | None] = mapped_column(String(10))
