+++ app/db/models/user_consent.py
@@ -1,17 +1,18 @@
 from datetime import datetime
 
-from sqlalchemy import TIMESTAMP, BigInteger, Boolean, ForeignKey, String, func
+from sqlalchemy import TIMESTAMP, Boolean, ForeignKey, String, func
 from sqlalchemy.orm import Mapped, mapped_column
 
 from app.db.base import Base
+from app.db.types import BIGINT
 
 
 class UserConsent(Base):
     __tablename__ = "user_consents"
 
-    consent_id: Mapped[int] = mapped_column(BigInteger, primary_key=True, autoincrement=True)
+    consent_id: Mapped[int] = mapped_column(BIGINT, primary_key=True, autoincrement=True)
     user_id: Mapped[int] = mapped_column(
-        BigInteger, ForeignKey("users.user_id", ondelete="CASCADE"), unique=True, nullable=False
+        BIGINT, ForeignKey("users.user_id", ondelete="CASCADE"), unique=True, nullable=False
     )
     terms_required: Mapped[bool] = mapped_column(Boolean, nullable=False, server_default="0")
     privacy_required: Mapped[bool] = mapped_column(Boolean, nullable=False, server_default="0")
